requirements.txt
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mega.py==1.0.8
pathlib==1.0.1
pycryptodome==3.21.0
python-telegram-bot==21.11.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tenacity==5.1.5
urllib3==2.3.0
